requirements.txt
alembic==1.1.0
aniso8601==8.0.0
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
dnspython==1.16.0
email-validator==1.0.4
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.23.0
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.15.6
